image_classification/requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
gensim
matplotlib
tqdm
torch
torchvision
numpy
pytorch-ignite
pandas